fix styles

+++ middleware/jwt.go
@@ -58,7 +58,7 @@ func JWT() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		email := c.Request.Header.Get("Remote-Email")
 		if len(email) == 0 {
-			c.Set("UserId", 0)
+			c.Set("UserId", uint(0))
 		} else {
 			c.Set("UserId", getId(email))
 		}
